chore: Move `server_config::routes` to a dedicated file
